Fix Generales - Navegacion - Seguridad

+++ app/Http/Controllers/PlaylistController.php
@@ -6,19 +6,18 @@ use App\Playlist;
 use App\Song;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Redirect;
+use UxWeb\SweetAlert\SweetAlert;
 
 class PlaylistController extends Controller
 {
     public function index(){
-
-        //$userId = Auth::id();
 
         $user = Auth::user();
 
         $playlists = Playlist::where('user_id',$user->id)->get();
 
         return view('playlist.index')->with('playlists', $playlists);
-
     }
 
     public function details($id) {
@@ -26,6 +25,14 @@ class PlaylistController extends Controller
         $songs = Song::where('playlist_id',$id)->get();
 
         $playlist = Playlist::find($id);
+
+        // Controles de seguridad.
+        $userId = Auth::id();
+
+        if($playlist->user_id != $userId){
+            $message = "No esta permitido ver playlists de otros usuarios";
+            return view('auth.error')->with('message',$message);
+        }
 
         return view('playlist.details')->with('songs', $songs)->with('playlist',$playlist);
     }
@@ -56,20 +63,58 @@ class PlaylistController extends Controller
             $playlist->save();
         }
 
+        $message = "Playlist creada con exito!";
+        $url = '/playlist/create';
+        $action = "Crear nueva Playlist";
+
+        $result = view('playlist.result-message')
+            ->with('message',$message)
+            ->with('url',$url)
+            ->with('action',$action)
+            ->render();
+
+        return response()->json(array('success' => true, 'html' => $result));
+
     }
 
     public function update($id){
+
         $playlist = Playlist::find($id);
+
+        // Controles de seguridad.
+        $userId = Auth::id();
+
+        if($playlist->user_id != $userId){
+            $message = "No esta permitido ver playlists de otros usuarios";
+            return view('auth.error')->with('message',$message);
+        }
 
         return view('playlist.create')->with('playlist',$playlist);
     }
 
     public function delete($id){
+
+        $userId = Auth::user()->id;
+
+        if($id != $userId){
+            $message = "No esta permitido modificar playlists de otros usuarios";
+            return view('auth.error')->with('message',$message);
+        }
+
         $playlist = Playlist::find($id);
         $playlist->delete();
+
     }
 
     public function changeVisibility(Request $request){
+
+
+        $userId = Auth::user()->id;
+
+        if($request->id != $userId){
+            $message = "No esta permitido modificar playlists de otros usuarios";
+            return view('auth.error')->with('message',$message);
+        }
 
         $playlist = Playlist::find($request->get('id'));
